+++ src/main/java/com/demodayapi/controller/ProjectControler.java
@@ -2,6 +2,7 @@ package com.demodayapi.controller;
 import com.demodayapi.enums.DemodayStatusEnum;
 import com.demodayapi.exceptions.ThereIsNotPeriodOfSubmissionException;
 import com.demodayapi.exceptions.UserAlredyHasProjectCreatedException;
+import com.demodayapi.exceptions.UserIsNotAdminException;
 import com.demodayapi.models.Demoday;
 import com.demodayapi.models.Project;
 import com.demodayapi.models.User;
@@ -66,7 +67,8 @@ public class ProjectControler {
 
 
      @DeleteMapping("/deleteprojects/{id}")
-        public ResponseEntity<Void> deleteProject(@PathVariable int id) {
+        public ResponseEntity<Void> deleteProject(@PathVariable int id,HttpServletRequest request) {
+        if(userService.isLoggedUserAdmin(request))throw new UserIsNotAdminException();
         projectService.deleteProjectById(id);
         return ResponseEntity.noContent().build();
     }
